Fix: 도로 모달 조회 시 위/경도 Double 타입 변환 및 쿼리 수정(#24)

+++ src/main/java/org/sesac/slopedbe/roadreport/service/RoadReportCenterService.java
@@ -1,6 +1,5 @@
 package org.sesac.slopedbe.roadreport.service;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -9,5 +8,5 @@ import org.sesac.slopedbe.roadreport.model.entity.RoadReportCenter;
 
 public interface RoadReportCenterService {
 	List<RoadReportCenterDTO> getComplaintCenterList();
-	Optional<RoadReportCenter> findClosestCenter(BigDecimal longitude, BigDecimal latitude, String mappingCity);
+	Optional<RoadReportCenter> findClosestCenter(Double longitude, Double latitude, String mappingCity);
 }
